Catch elastic search exceptions as failed futures
ES queries can throw exceptions before the ListenableActionFuture is
created.  Kaka’s scheduler stops repeating if the runnable throws an
exception.  This combination means that if the cluster is unreachable,
the healthcheck fails and then never runs again.

Catching this early exception and wrapping it as a failed future should
fix this.